Add `CONTRIBUTING.md`
Including instructions for new contributors to open source is a courtesy that will foster a welcoming and safe open source community at Facebook. Even though this project is not currently very active, it makes sense to include this document.

We link to the `CODE_OF_CONDUCT.md` here, and a follow up commit will add that.